Modifies l2r script

+++ l2r_train_eval.py
@@ -1,3 +1,4 @@
+import torch
 from collections import defaultdict
 from dataclasses import dataclass
 import numpy as np
@@ -11,6 +12,9 @@ from config import TrainConfig
 
 from train import train_stage1, train_stage2, eval_stage1, eval_stage2
 
+L2R_TRAIN_TIME_DATA = ["image", "mask"]
+L2R_TEST_TIME_DATA = ["label", "keypoints"]
+
 @dataclass
 class SomeNetCheckpoint:
     stage: int
@@ -22,7 +26,7 @@ def find_last_checkpoint(config: TrainConfig, checkpointroot: Path) -> SomeNetCh
     Find the last checkpoint
 
     NOTE: if the config changes, i.e. previous stage number of steps has changed
-    this function won't spot it. Possible to refactor this function and retain
+    this function won't spot it. Possible to implement this function and retain
     old behaviour with a flag, but its probably better to avoid the situation above
     entirely by specifying a different save path.
     """
@@ -33,9 +37,14 @@ def find_last_checkpoint(config: TrainConfig, checkpointroot: Path) -> SomeNetCh
         if stage_folder.exists():
             checkpoints = [f.name for f in stage_folder.iterdir() if f.name.endswith('.pth')]
             if len(checkpoints) == 0:
-                # NOTE: you cannot be sure if artifacts for the last stage were fully generated
-                # so we go the safe route and go back to the previous stage
-                continue
+                # NOTE: we assume that the checkpoints from the previous stage were generated
+                return SomeNetCheckpoint(
+                    stage=stage,
+                    step=0,
+                    checkpoint=None
+                )
+
+                # continue
             last_checkpoint = max(checkpoints, key=get_step)
             return SomeNetCheckpoint(
                 stage=stage,
@@ -62,11 +71,11 @@ def get_split_pairs_from_paired_dataset(data: Dict, root: Path):
 
     val_fixeds = [dat["fixed"] for dat in data["registration_val"]]
 
-    for split_p in ("training", "test"):
+    for split_p in ["training"]:
         for dat in data[f"{split_p}_paired_images"]:
             split = "val" if dat["fixed"] in val_fixeds else "train"
             pair_dat = {}
-            for lab in ("fixed", "moving"):
+            for lab in ["fixed", "moving"]:
                 image = dat[lab]
                 if image in train_data:
                     all_dat_for_image = train_data[image]
@@ -74,26 +83,33 @@ def get_split_pairs_from_paired_dataset(data: Dict, root: Path):
                         {f"{lab}_{k}": v for k, v in all_dat_for_image.items()}
                     )
                 else:
-                    print(f"{image} not found")
-                    break
+                    raise ValueError(f"{image} not found")
             else:
                 split_data[split].append(pair_dat)
+
+    # FIXME: their test specification is a mess
+    # for dat in data[f"test_paired_images"]:
+    #     pair_dat = {}
+    #     for lab in ("fixed", "moving"):
+    #         image = dat[lab]
+    #         if image in train_data:
+    #             all_dat_for_image = train_data[image]
+    #             pair_dat.update(
+    #                 {f"{lab}_{k}": v for k, v in all_dat_for_image.items()}
+    #             )
+    #         else:
+    #             raise ValueError(f"{image} not found")
+    #     else:
+    #         split_data["test"].append(pair_dat)
 
     return split_data
 
 
 def get_split_pairs_from_unpaired_dataset(data: Dict, root: Path):
     train_data = {}
-    test_data = {}
     split_data = defaultdict(list)
     for dat in data["training"]:
         train_data[dat["image"]] = {
-            (k if k != "label" else "segmentation"): str(root / v)
-            for k, v in dat.items()
-        }
-
-    for dat in data["test"]:
-        test_data[dat["image"]] = {
             (k if k != "label" else "segmentation"): str(root / v)
             for k, v in dat.items()
         }
@@ -120,20 +136,11 @@ def get_split_pairs_from_unpaired_dataset(data: Dict, root: Path):
         pair_dat = {}
         for lab in ("fixed", "moving"):
             image = dat[lab]
-            if image in train_data:
-                all_dat_for_image = train_data[image]
-                pair_dat.update({f"{lab}_{k}": v for k, v in all_dat_for_image.items()})
-            else:
-                print(f"{image} not found")
-                break
-        else:
-            split_data["test"].append(pair_dat)
 
     return split_data
 
-
 def get_pairs_from_list(
-    data: List[Dict[str, Path]], pairs_per_image: int = 5, root: Optional[Path] = None
+    data: List[Dict[str, Path]], pairs_per_image: int = 1, root: Optional[Path] = None
 ) -> List[Dict[str, Path]]:
     return get_random_pairs_from_list(data, pairs_per_image)
 
@@ -237,6 +244,8 @@ def main(dataset_json: Path, config_json: Path):
         config_dict = json.load(f)
         config = TrainConfig(**config_dict)
 
+    if config.gpu_num is not None:
+        torch.cuda.set_device(config.gpu_num)
     root = dataset_json.parent
     metadata = {}
     training_recipe = {}
@@ -275,11 +284,11 @@ def main(dataset_json: Path, config_json: Path):
         config.savedir.mkdir(exist_ok=True)
         checkpointroot = config.savedir / Path(f"checkpoints")
 
-    resumption_point = None
+    resumption_point: Optional[SomeNetCheckpoint] = None
     if checkpointroot.exists() and (checkpointroot/"stage1").exists() and not config.overwrite:
         try:
             resumption_point = find_last_checkpoint(config, checkpointroot)
-            print(f"Resuming from stage {resumption_point.stage} step {resumption_point.step}")
+            print(f"Resuming from stage {resumption_point.stage-1} step {resumption_point.step}")
         except ValueError:
             resumption_point = None
 
@@ -287,10 +296,10 @@ def main(dataset_json: Path, config_json: Path):
 
     data_json = checkpointroot / f"data.json"
 
-    with open(data_json, "w") as f:
-        json.dump(split_pairs, f)
+    if not data_json.exists():
+        with open(data_json, "w") as f:
+            json.dump(split_pairs, f)
 
-    last_checkpoint = None
     start = 0 if resumption_point is None else resumption_point.stage-1
     for i, stage in enumerate(config.stages[start:], start=start):
         if i == 0:
@@ -308,8 +317,13 @@ def main(dataset_json: Path, config_json: Path):
                 log_freq=stage.log_freq,
                 val_freq=stage.val_freq,
                 start=None if resumption_point is None else resumption_point.checkpoint,
-                starting_step=None if resumption_point is None else resumption_point.step
+                starting_step=None if resumption_point is None else resumption_point.step,
+                noisy=config.noisy,
+                noisy_v2=config.noisy_v2,
+                image_loss_fn=config.image_loss_fn,
+                image_loss_weight=config.image_loss_weight
             )
+            resumption_point = None
 
             last_checkpoint = (checkpointroot
                     / f"stage{i+1}"
@@ -329,12 +343,6 @@ def main(dataset_json: Path, config_json: Path):
                     split=split,
                 )
         else:
-            if resumption_point is not None:
-                starting_checkpoint = resumption_point.checkpoint
-            elif stage.start_from_last:
-                starting_checkpoint = last_checkpoint
-            else:
-                starting_checkpoint = None
             print(f"Training stage {i}")
             train_stage2(
                 data_json=data_json,
@@ -346,12 +354,18 @@ def main(dataset_json: Path, config_json: Path):
                 iters=stage.iters,
                 search_range=stage.search_range,
                 diffeomorphic=config.diffeomorphic,
-                start=starting_checkpoint,
+                start=None if resumption_point is None else resumption_point.checkpoint,
                 save_freq=stage.save_freq,
                 log_freq=stage.log_freq,
                 val_freq=stage.val_freq,
-                starting_step=None if resumption_point is None else resumption_point.step
+                starting_step=None if resumption_point is None else resumption_point.step,
+                noisy=config.noisy,
+                noisy_v2=config.noisy_v2,
+                image_loss_fn=config.image_loss_fn,
+                image_loss_weight=config.image_loss_weight
             )
+
+            resumption_point = None
 
             last_checkpoint = (checkpointroot
                     / f"stage{i+1}"
